cleaned up project

+++ Frameworks/cocos2d-iphone-classic-2.2/cocos2d/CCLayer.m
@@ -714,7 +714,7 @@
 
 	// Compressed Interpolation mode
 	if( _compressedInterpolation ) {
-		float h2 = 1 / ( fabsf(u.x) + fabsf(u.y) );
+		float h2 = 1 / ( fabs(u.x) + fabs(u.y) );
 		u = ccpMult(u, h2 * (float)c);
 	}
